+++ TangramKit/TGLinearLayout.swift
@@ -335,7 +335,7 @@ open class TGLinearLayout: TGBaseLayout,TGLinearLayoutViewSizeClass {
         
         
         tgAdjustLayoutSelfSize(selfSize: &selfSize, lsc: lsc)
-        
+        tgAdjustSubviewsLayoutTransform(sbs: sbs, lsc: lsc, selfSize: selfSize)
         tgAdjustSubviewsRTLPos(sbs: sbs, selfWidth: selfSize.width)
         
         return self.tgAdjustSizeWhenNoSubviews(size: selfSize, sbs: sbs, lsc:lsc)