+++ TangramKit/TangramKit.swift
@@ -53,7 +53,7 @@
  
  */
 
-//Current version is 1.1.6, please open: https://github.com/youngsoft/TangramKit/blob/master/CHANGELOG.md to show the changes.
+//Current version is 1.2.0, please open: https://github.com/youngsoft/TangramKit/blob/master/CHANGELOG.md to show the changes.